Split up addElementToGame method
